backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pymilvus>=2.3.0
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cu118
torchvision>=0.15.0 --index-url https://download.pytorch.org/whl/cu118
numpy>=1.24.0
pillow>=9.5.0
open_clip_torch>=2.20.0
requests>=2.28.0
python-multipart>=0.0.6
scikit-learn
aiohttp
python-dotenv
websockets
regex
ftfy
tqdm
